feat: adjust input size, refactor code
