+++ public/i18n/en.yaml
@@ -779,7 +779,7 @@ kyma-companion:
     http-error: Response status code is {{statusCode}}. Retrying {{attempt}}.
   introduction: Hi, I am your Kyma assistant! You can ask me any question, and I will try to help you to the best of my abilities. Meanwhile, you can check the suggested questions below; you may find them helpful!
   introduction-no-suggestions: Hi, I am your Kyma assistant! You can ask me any question, and I will try to help you to the best of my abilities. While I don't have any initial suggestions for this resource, feel free to ask me anything you'd like!
-  placeholder: Message Joule
+  placeholder: Message Joule...
   disclaimer: Joule is powered by generative AI, and the output must be reviewed before use. Do not enter any sensitive personal data, and avoid entering any other data you do not wish to be processed.
   tabs:
     chat: Chat
